API server/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sqlalchemy[asyncio]
asyncpg
greenlet
pydantic-settings
alembic
httpx
loguru
openpyxl
redis>=5.0.0
pandas
webpush==1.0.5
py-vapid
aiohttp
cryptography
passlib[bcrypt]
PyJWT
requests
python-multipart
user-agents
geoip2
maxminddb